[swdb]: reintroduce trans performed with feature
TRANS_WITH table was missing in SWDB design and history info output.

History info should be fine now.
