+++ libdnf/CMakeLists.txt
@@ -34,7 +34,8 @@ SET (LIBDNF_SRCS
     dnf-swdb.c
     dnf-swdb-db.c
     dnf-swdb-groups.c
-    dnf-swdb-trans.c)
+    dnf-swdb-trans.c
+    dnf-swdb-pkg.c)
 
 configure_file ("dnf-version.h.in"  ${CMAKE_CURRENT_SOURCE_DIR}/dnf-version.h)
 configure_file ("libdnf.pc.in" ${CMAKE_CURRENT_BINARY_DIR}/libdnf.pc @ONLY)
@@ -117,6 +118,9 @@ SET(LIBDNF_headers
     dnf-swdb-groups.h
     dnf-swdb-groups-sql.h
     dnf-swdb-trans.h
+    dnf-swdb-trans-sql.h
+    dnf-swdb-pkg.h
+    dnf-swdb-pkg-sql.h
     libdnf.h
     )
 
@@ -175,6 +179,7 @@ if (GOBJECT_INTROSPECTION_FOUND)
                     dnf-swdb-groups.c dnf-swdb-groups.h
                     dnf-swdb-db.c dnf-swdb-db.h
                     dnf-swdb-trans.c dnf-swdb-trans.h
+                    dnf-swdb-pkg.c dnf-swdb-pkg.h
             DEPENDS libdnf
             WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
