+++ tests/libdnf/CMakeLists.txt
@@ -9,6 +9,8 @@ SET(SWDB_SOURCES
     ${CMAKE_SOURCE_DIR}/libdnf/dnf-swdb-groups.h
     ${CMAKE_SOURCE_DIR}/libdnf/dnf-swdb-trans.c
     ${CMAKE_SOURCE_DIR}/libdnf/dnf-swdb-trans.h
+    ${CMAKE_SOURCE_DIR}/libdnf/dnf-swdb-pkg.c
+    ${CMAKE_SOURCE_DIR}/libdnf/dnf-swdb-pkg.h
     )
 
 ADD_EXECUTABLE(test_libdnf_swdb dnf-swdb-test.c ${SWDB_SOURCES})
